requirements.txt
arabic-reshaper==2.1.3
asgiref==3.5.1
backports.entry-points-selectable==1.1.0
barcode==1.0.2
beautifulsoup4==4.10.0
certifi==2021.10.8
charset-normalizer==2.0.7
defusedxml==0.7.1
diff-match-patch==20200713
distlib==0.3.2
Django==4.0.4
django-auto-logout==0.5.0
django-bootstrap-modal-forms==2.2.0
django-bootstrap3==21.2
django-bootstrap5==21.3
django-ckeditor==6.4.0
django-crispy-forms==1.13.0
django-import-export==2.8.0
django-js-asset==2.0.0
django-phone-field==1.8.1
django-widget-tweaks==1.4.8
django-xhtml2pdf==0.0.4
et-xmlfile==1.1.0
filelock==3.0.12
future==0.18.2
GDAL @ file:///C:/Users/HP/Downloads/GDAL-3.3.2-cp39-cp39-win_amd64.whl
html5lib==1.1
idna==3.3
mailchimp-marketing==3.0.70
MarkupPy==1.14
mysqlclient==2.1.0
odfpy==1.4.1
openpyxl==3.0.9
paywix==1.4.0
Pillow==8.4.0
platformdirs==2.3.0
psycopg2==2.9.1
PyJWT==2.3.0
PyMySQL==1.0.2
PyPDF2==1.26.0
python-barcode==0.13.1
python-bidi==0.4.2
python-dateutil==2.8.2
python-Levenshtein==0.12.2
pytz==2022.1
PyYAML==6.0
reportlab==3.6.2
requests==2.26.0
sendsms==0.2.0
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
tablib==3.2.1
twilio==7.2.0
tzdata==2022.1
urllib3==1.26.7
virtualenv==20.7.2
webencodings==0.5.1
xhtml2pdf==0.2.5
xlrd==2.0.1
xlwt==1.3.0
